patch native function arg parsing
