+++ language/diem-framework/modules/0L/TowerState.move
@@ -21,7 +21,7 @@ module TowerState {
     use 0x1::VDF;
     use 0x1::Vector;
 
-    const VDF_SECURITY_PARAM: u64 = 2048;
+    use 0x1::Debug::print;
 
     const EPOCHS_UNTIL_ACCOUNT_CREATION: u64 = 6;
 
@@ -176,6 +176,7 @@ module TowerState {
       difficulty: u64,
       security: u64,
     ) acquires TowerProofHistory, TowerList, TowerStats {
+      print(&001);
       // TODO: Previously in OLv3 is_genesis() returned true. 
       // How to check that this is part of genesis? is_genesis returns false here.
 
@@ -266,7 +267,7 @@ module TowerState {
       steady_state: bool
     ) acquires TowerProofHistory, TowerList, TowerStats {
       let miner_history = borrow_global<TowerProofHistory>(miner_addr);
-
+      print(&10010);
       let epoch = DiemConfig::get_current_epoch();
       if (epoch < 60) { // network is bootstrapping
         assert(
@@ -280,26 +281,30 @@ module TowerState {
         );
       };
 
-
+      print(&10020);
       // If not genesis proof, check hash to ensure the proof continues the chain
       if (steady_state) {
         //If not genesis proof, check hash 
         assert(&proof.challenge == &miner_history.previous_proof_hash,
         Errors::invalid_state(130107));      
       };
+      print(&10030);
 
-      let valid = VDF::verify(&proof.challenge, &proof.difficulty, &proof.solution, &VDF_SECURITY_PARAM);
+      let valid = VDF::verify(&proof.challenge, &proof.difficulty, &proof.solution, &proof.security);
       assert(valid, Errors::invalid_argument(130108));
+      print(&10040);
 
       // add the miner to the miner list if not present
       increment_miners_list(miner_addr);
 
+      print(&10050);
       // Get a mutable ref to the current state
       let miner_history = borrow_global_mut<TowerProofHistory>(miner_addr);
 
       // update the miner proof history (result is used as seed for next proof)
       miner_history.previous_proof_hash = Hash::sha3_256(*&proof.solution);
-      
+        print(&10050);
+    
       // Increment the verified_tower_height
       if (steady_state) {
         miner_history.verified_tower_height = miner_history.verified_tower_height + 1;
@@ -308,10 +313,14 @@ module TowerState {
         miner_history.verified_tower_height = 0;
         miner_history.count_proofs_in_epoch = 1
       };
-    
+      print(&10060);
+
       miner_history.latest_epoch_mining = DiemConfig::get_current_epoch();
+      print(&10070);
 
       increment_stats(miner_addr);
+      print(&10080);
+
     }
 
     // Checks that the validator has been mining above the count threshold
@@ -399,7 +408,13 @@ module TowerState {
     // Function to initialize miner state
     // Permissions: PUBLIC, Signer, Validator only
     // Function code: 07
-    public fun init_miner_state(miner_sig: &signer, challenge: &vector<u8>, solution: &vector<u8>, difficulty: u64, security: u64) acquires TowerProofHistory, TowerList, TowerStats {
+    public fun init_miner_state(
+      miner_sig: &signer,
+      challenge: &vector<u8>,
+      solution: &vector<u8>,
+      difficulty: u64,
+      security: u64
+    ) acquires TowerProofHistory, TowerList, TowerStats {
       
       // NOTE Only Signer can update own state.
       // Should only happen once.
@@ -425,7 +440,7 @@ module TowerState {
         solution: *solution,
         security,
       };
-
+      print(&10000);
       //submit the proof
       verify_and_update_state(Signer::address_of(miner_sig), proof, false);
     }
