+++ language/move-lang/functional-tests/tests/0L/vdf/vdf_verify_valid_proof.move
@@ -12,7 +12,8 @@ fun main() {
   // Generate solutions with cd ./verfiable_delay/vdf-cli && cargo run -- -l=2048 aa 100
   // the -l=2048 is important because this is the security paramater of 0L miner.
   let solution = TestFixtures::easy_sol();
+  let security = 2048;
 
-  assert(VDF::verify(&challenge, &difficulty, &solution) == true, 1);
+  assert(VDF::verify(&challenge, &difficulty, &solution, &security) == true, 1);
 }
 }
