--- ol/tower/test.prove_5m_256
@@ -1,1 +1,0 @@
-0008bc09e6166510a94621cc41b3751780fffc31e659e3ea5ad5996c8bea13ebe46900322198e8a6e67e983df51ce82d684310001ac44ef7cfef8e66f54eab5b9a87bff7
