--- ol/tower/test.prove_5m_512
@@ -1,1 +1,0 @@
-004cceceab990902f8091d5a5f862cb572e6a849b997f8cb14e01868508693021800478c2fab72d55b5577f34890f03e9cdfe47771724327f0c112dd6a1dfaaefed30068526590a7bfbef7c0a1d6320f0ecbb2d5bdc55952bfddbf166f5a0b7ea48791003eeff4492530be6e9b6e98bc24ea8411dedff40ff918195c87c6c85b1959331b
